# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Safalife API is a comprehensive backend service for Safalife's competition management system. This API provides endpoints for health monitoring, user management, competition tracking, and administrative functions. Built with Go using Clean Architecture principles, it offers high performance and reliability for managing large-scale competitions.",
        "title": "Safalife API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Safalife API Support Team",
            "url": "http://www.safalife.com/support",
            "email": "api-support@safalife.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/quran/bookmarks/ayahs/{id}": {
            "get": {
                "description": "Get a specific bookmark by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Get bookmark ayah by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "ayah"
                        ],
                        "type": "string",
                        "example": "ayah",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.BookmarkAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Bookmark not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a bookmark by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Delete bookmark ayah",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bookmark ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookmark deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteBookmarkAyahSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Bookmark not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/last-reads/{id}": {
            "get": {
                "description": "Get a specific last read entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Get last read by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LastRead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "relations"
                        ],
                        "type": "string",
                        "example": "relations",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update fields of an existing last read entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Update last read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LastRead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "LastRead update information",
                        "name": "lastRead",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLastReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read updated successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.LastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a last read entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LastRead"
                ],
                "summary": "Delete last read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LastRead ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Last read deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteLastReadSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Last read not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/quran/progress-hatam/{id}": {
            "get": {
                "description": "Get a specific progress hatam entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProgressHatam"
                ],
                "summary": "Get progress hatam by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProgressHatam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "relations"
                        ],
                        "type": "string",
                        "example": "relations",
                        "description": "Include related data",
                        "name": "include",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progress hatam retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.ProgressHatamSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Progress hatam not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update fields of an existing progress hatam entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProgressHatam"
                ],
                "summary": "Update progress hatam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProgressHatam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ProgressHatam update information",
                        "name": "progressHatam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateProgressHatamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progress hatam updated successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.ProgressHatamSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Progress hatam not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a progress hatam entry by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProgressHatam"
                ],
                "summary": "Delete progress hatam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ProgressHatam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Progress hatam deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/quran.DeleteProgressHatamSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Progress hatam not found",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/quran.QuranErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "dto.BookmarkAyahResponse": {
            "type": "object",
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.LastReadResponse": {
            "type": "object",
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "ayahNumber": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lastReadAt": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number"
                },
                "surahId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.ProgressHatamResponse": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "juzId": {
                    "type": "string"
                },
                "lastAyahId": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number"
                },
                "startAyahId": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateLastReadRequest": {
            "type": "object",
            "required": [
                "ayahId",
                "ayahNumber",
                "id",
                "progressPct",
                "updatedBy"
            ],
            "properties": {
                "ayahId": {
                    "type": "string"
                },
                "ayahNumber": {
                    "type": "integer",
                    "minimum": 1
                },
                "id": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "updatedBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
        "dto.UpdateProgressHatamRequest": {
            "type": "object",
            "required": [
                "id",
                "isCompleted",
                "lastAyahId",
                "progressPct",
                "updatedBy"
            ],
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isCompleted": {
                    "type": "boolean"
                },
                "lastAyahId": {
                    "type": "string"
                },
                "progressPct": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                },
                "updatedBy": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "quran.BookmarkAyahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.BookmarkAyahResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Bookmark retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.DeleteBookmarkAyahSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Bookmark deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.DeleteLastReadSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Last read deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.DeleteProgressHatamSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "null"
                },
                "message": {
                    "type": "string",
                    "example": "Progress hatam deleted successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.LastReadSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.LastReadResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Last read retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
        "quran.ProgressHatamSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ProgressHatamResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Progress hatam retrieved successfully"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "quran.QuranErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "$ref": "#/definitions/core.ErrorDetail"
                },
                "message": {
                    "type": "string",
                    "example": "Failed to process request"
                },
                "responseTime": {
                    "type": "string",
                    "example": "15.234ms"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}